requirements.txt
google-api-python-client==2.97.0
google-auth==2.23.0
google-auth-oauthlib==1.0.0
anthropic==0.3.11
supabase==1.0.3
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
requests==2.31.0
pydantic>=2.1.0
streamlit==1.30.0
# Vector database and embedding dependencies
chromadb==0.4.23
sentence-transformers==2.5.1
faiss-cpu==1.7.4
langchain==0.1.5
langchain-text-splitters==0.0.1
langchain-community==0.0.17
diskcache==5.6.3
